dubiously tweak issue sorting to match my-jira-tickets
x.fields.status.id is the dubious part, since it is a string, not an
int.  It happens to sort the way we want, actually more so than if it
is converted to int - which tells me maybe it ought to be handled some
other way, like explicitly mapping statuses to our own ordering.

But i guess this suffices for the time being.