whew, make a non-auto-viv version of easydict
and rename the old auto-viv version easydict_av

notably the non-auto-viv version uses self.get(name), so it returns
None on missing keys instead of raising a key error

...

you can still assign values with 'e.xyz = x', just you can't do
'e.abc.xyz = x' if 'e.abc' doesn't exist.

the motivation here, by the way, is that most of the time we only
care about checking values (eg, params.issue), and while the default
auto-vivified thing (an empty easydict) evaluates to False in bool
context (which we want), it also sets issue={} which is not quite
what we want (and maybe even breaks things if we are not careful).